+++ build.sh
@@ -34,26 +34,19 @@ if [ "$JDK" != "" ]; then
   PROPERTIES="$PROPERTIES -Djdk8.home=/usr/lib/jvm/java-8-oracle"
 fi
 
-if [ "$TRAVIS_REPO_SLUG" == "Blazebit/blaze-persistence" ] && 
-    [ "$TRAVIS_BRANCH" == "master" ] &&
-    [ "$JPAPROVIDER" == "hibernate-5.2" || "$JPAPROVIDER" == "hibernate-apt" ] &&
+if [ "$JPAPROVIDER" == "hibernate-5.2" || "$JPAPROVIDER" == "hibernate-apt" ] &&
     [ "$RDBMS" == "h2" ]; then
   exec mvn -B -P ${JPAPROVIDER},${RDBMS},${SPRING_DATA:-spring-data-1.11.x},${DELTASPIKE:-deltaspike-1.7} clean install -V $PROPERTIES
 else
   PROJECT_LIST="core/testsuite,entity-view/testsuite,jpa-criteria/testsuite,integration/deltaspike-data/testsuite,integration/jaxrs,integration/spring-data/testsuite/webflux,integration/spring-data/testsuite/webmvc,examples/spring-data-webmvc,examples/spring-data-webflux,examples/showcase/runner/spring,examples/showcase/runner/cdi"
-  if [ "$TRAVIS_REPO_SLUG" == "Blazebit/blaze-persistence" ] &&
-    [ "$TRAVIS_BRANCH" == "master" ] &&
-    [ "$TRAVIS_PULL_REQUEST" == "false" ] &&
-    [ "$JPAPROVIDER" == "hibernate-6.0" ] &&
+  if [ "$JPAPROVIDER" == "hibernate-6.0" ] &&
     [ "$RDBMS" == "h2" ]; then
     # Just in case we want to run against a specific version
     #git clone --depth=1 --branch="wip/6.0" https://github.com/sebersole/hibernate-core.git hibernate6
     #cd hibernate6
     #./gradlew clean build publishToMavenLocal -x :documentation:buildDocs -x :documentation:aggregateJavadocs -x test -x findbugsMain -x findbugsTest -x checkStyleMain -x checkStyleTest
     : # do nothing right now
-  elif [ "$TRAVIS_REPO_SLUG" == "Blazebit/blaze-persistence" ] &&
-    [ "$TRAVIS_BRANCH" == "master" ] &&
-    [ "$JPAPROVIDER" == "hibernate-5.4" ]; then
+  elif [ "$JPAPROVIDER" == "hibernate-5.4" ]; then
     PROJECT_LIST="$PROJECT_LIST,integration/quarkus/deployment,examples/quarkus/testsuite/base"
     if [ "$NATIVE" == "true" ]; then
       if [ "$RDBMS" == "mysql8" ]; then
@@ -61,11 +54,11 @@ else
       else
         PROJECT_LIST="$PROJECT_LIST,examples/quarkus/testsuite/native/$RDBMS"
       fi
-      eval exec travis-pls mvn -B -P ${JPAPROVIDER},${RDBMS},${SPRING_DATA:-spring-data-1.11.x},${DELTASPIKE:-deltaspike-1.7},native clean install --projects $PROJECT_LIST -am -V $PROPERTIES
+      exec travis-pls mvn -B -P ${JPAPROVIDER},${RDBMS},${SPRING_DATA:-spring-data-1.11.x},${DELTASPIKE:-deltaspike-1.7},native clean install --projects $PROJECT_LIST -am -V $PROPERTIES
     else
-      eval exec mvn -B -P ${JPAPROVIDER},${RDBMS},${SPRING_DATA:-spring-data-1.11.x},${DELTASPIKE:-deltaspike-1.7} clean install --projects $PROJECT_LIST -am -V $PROPERTIES
+      exec mvn -B -P ${JPAPROVIDER},${RDBMS},${SPRING_DATA:-spring-data-1.11.x},${DELTASPIKE:-deltaspike-1.7} clean install --projects $PROJECT_LIST -am -V $PROPERTIES
     fi
   else
-    eval exec mvn -B -P ${JPAPROVIDER},${RDBMS},${SPRING_DATA:-spring-data-1.11.x},${DELTASPIKE:-deltaspike-1.7} clean install --projects $PROJECT_LIST -am -V $PROPERTIES
+    exec mvn -B -P ${JPAPROVIDER},${RDBMS},${SPRING_DATA:-spring-data-1.11.x},${DELTASPIKE:-deltaspike-1.7} clean install --projects $PROJECT_LIST -am -V $PROPERTIES
   fi
 fi